Make the purpose of database_observability.ExtractExplainPlanOutputFromLogMsg more apparent
